Refactor: Gas tile overlay split (#42881)
* yeah

* split2

* well

* cleanup

* docs fix

* Dependency cleanup

* Update Content.Client/Atmos/EntitySystems/GasTileFireOverlaySystem.cs


* Update Content.Client/Atmos/Overlays/GasTileVisibleGasOverlay.cs


* Update Content.Client/Atmos/Overlays/GasTileVisibleGasOverlay.cs


* yup

* modifying a lot anyways

* other cleanup

---------
